requirements.txt
django
django_rest_framework
djangorestframework-simplejwt
drf-yasg
djoser